Refactor into package
